Add Extensions and PerMessageCompressionContext
Add types representing the configuration for, and then actual in-use
subset of, websocket protocol extensions for a connection. Currently
only permessage-deflate is supported, but having an enum is zero-cost
when it has only zero or one variants, and an empty struct takes up no
space.

This also lets us reduce the amount of code that needs to be annotated
with `#[cfg(feature = "deflate")]` since calling code can always assume
that the PerMessageCompressionContext type exists even if its method
impls are sometimes no-ops.

Based on work by Benjamin Swart <Benjaminswart@email.cz>